bot/requirements.txt
aiogram==3.13.1
python-dotenv==1.2.2
httpx==0.28.1

